Support parsing json in addition to loading it
Create a seperate subcommand for `dump_json` which converts
the XML into a JSON dump.
Based on my tests it takes about ~24 seconds to parse XML
while it only takes about ~15 second to parse JSON.
All of this is in debug mode of course,
so it should be much faster in release.